Refactored code into a separate file.
